requirements.txt
streamlit
google-cloud-aiplatform
google-generativeai
pandas
seaborn
matplotlib
google-auth-oauthlib